+++ packages/server/lib/privileged-commands/privileged-commands-manager.ts
@@ -7,6 +7,7 @@ import * as exec from '../exec'
 import * as files from '../files'
 import { fs } from '../util/fs'
 import task from '../task'
+import { getUserDefinedEnvironmentVariables } from '../getUserDefinedEnvs'
 
 export interface SpecChannelOptions {
   isSpecBridge: boolean
@@ -102,6 +103,8 @@ class PrivilegedCommandsManager {
     }
 
     switch (commandName) {
+      case 'env':
+        return getUserDefinedEnvironmentVariables({ requestedKeys: options.envVars, userDefinedEnvironmentVariables: config.env || {} })
       case 'exec':
         return exec.run(config.projectRoot, options)
       case 'origin':