+++ system-tests/projects/disallow-cypress-env/cypress-without-disallow-cypress-env.config.ts
@@ -1,0 +1,12 @@
+import { defineConfig } from 'cypress'
+
+export default defineConfig({
+  disallowCypressEnv: false,
+  e2e: {
+    supportFile: false,
+    fixturesFolder: false,
+    setupNodeEvents (on, config) {
+      // implement node event listeners here
+    },
+  },
+})